Fix mypy errors in batch 2 part 3 (#1282)
plane_wave_hamiltonian: annotate spins as list[Optional[int]]. physical_costing: MagicStateFactory rounds and failure_rate are floats at runtime.

+++ src/openfermion/resource_estimates/surface_code_compilation/physical_costing.py
@@ -9,8 +9,8 @@ from typing import Tuple, Iterator
 class MagicStateFactory:
     details: str
     physical_qubit_footprint: int
-    rounds: int
-    failure_rate: int
+    rounds: float
+    failure_rate: float
 
 
 @dataclasses.dataclass(frozen=True, unsafe_hash=True)
